requirements.txt
celery
psycopg2-binary
redis